Commented-out some test code in SentinelLdap.php.
Secured Cartalyst package URL in composer.json.

Changed calls to `register()` to `boot()` because I just learned that `register()` is called regardless of an extension's status (Enabled / Disabled), whereas `boot()` is only called for Enabled extensions. This change was necessary to be able to completely toggle LDAP functionality on and off using Platform's GUI.