Moved facades to a more appropriate location (they should be inside the /src directory).
Updated composer.json per above change, and combined the two database paths into a single path (of which both are subdirectories).

Eliminated use of App facade (since $this->app is equivalent), made a similar change for a wasteful instantiation of the Sentinel class, eliminated unused namespace references, and added a facade alias for MmicLdap.