fix: Remove duplicate function call that causes exception
In PHP 8, the duplicate function call throws "ldap_close(): supplied resource is not a valid ldap link resource", because the call to ldap_unbind(), changes $conn from an `ldap link resource` to a `Closed resource`.

ldap_close() is merely an alias of ldap_unbind(), so calling the former after the latter is unnecessary.